requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
beautifulsoup4==4.13.3
Brotli==1.0.9
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==2.1.0
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
cssselect2==0.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.7
django-admin-interface==0.29.4
django-ckeditor==6.7.2
django-ckeditor-5==0.2.17
django-colorfield==0.12.0
django-jazzmin==3.0.1
django-js-asset==3.1.2
fastapi==0.115.12
ffmpeg-python==0.2.0
filelock==3.18.0
fsspec==2025.3.2
future==1.0.0
git-filter-repo==2.47.0
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
huggingface-hub==0.30.1
hyperframe==6.0.0
idna==3.4
imageio==2.37.0
imageio-ffmpeg==0.6.0
intel-cmplr-lib-ur==2025.1.0
intel-openmp==2025.1.0
Jinja2==3.1.6
llvmlite==0.44.0
lxml==5.3.2
MarkupSafe==3.0.2
mkl==2025.1.0
mkl-service==2.4.2
mkl_fft==1.3.13
mkl_random==1.2.10
more-itertools==10.6.0
moviepy==2.1.2
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
openai-whisper==20240930
oscrypto==1.3.0
packaging==24.2
pillow==10.4.0
proglog==0.1.11
protobuf==6.30.2
psutil==7.0.0
pycparser==2.21
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
pypdf==5.4.0
PySocks @ file:///D:/bld/pysocks_1733217287171/work
python-bidi==0.6.6
python-dotenv==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==8.1
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
safetensors==0.5.3
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
starlette==0.46.1
svglib==1.5.1
sympy==1.13.1
tbb==2022.1.0
tcmlib==1.3.0
text-unidecode==1.3
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
umf==0.10.0
uritools==4.0.3
urllib3==1.26.14
uvicorn==0.34.0
webencodings==0.5.1
whitenoise==6.9.0
win_inet_pton==1.0.0
xhtml2pdf==0.2.17
yt-dlp==2025.3.31
zstandard==0.23.0
